- add session clearing interface and method, deprecate old import locations
